FAT-6 update Depends
