+++ src/schemas/model.py
@@ -1,5 +1,10 @@
+from typing import Optional
+
+from fastapi import Query
 from pydantic import BaseModel
 from uuid import UUID
+
+from schemas.enums import OrderBy
 
 
 class Author(BaseModel):
@@ -57,3 +62,12 @@ class PostUpdate(BaseModel):
     model_config = {
         "from_attributes": True
     }
+
+
+# Модель фильтров
+class PostFilters(BaseModel):
+    limit: int = Query(20, ge=5, le=100)
+    offset: int = Query(0, ge=0)
+    search: Optional[str] = Query(None, min_length=1)
+    order_by: OrderBy = Query(OrderBy.CREATED_AT)
+    descending: bool = Query(False)
